Making the gpg-agent run by root.
It was difficult to use it as a separate user.
The standard socket is used by the preseting command,
and the extra socket is the one used by the proxy code.
Requests are sent to the extra socket, so that workers have
a limited range of possibilities on the agent.