Change faulty config variable
In the documentation it states that the variable to configure the s3 keys should be `VAULT_ACCESS_KEY` and `VAULT_SECRET_KEY` respectively.
Yet in the code there is a dependency on a environmental variables named `S3_ACCESS_KEY` and `S3_SECRET_KEY`. These changes removes the `S3` prefixed entries and replaces them with the `VAULT` prefixed entries.