docs(CONTRIB): readthedocs renders the contributing guidelines
The changes to the conf.py is to make sure that Sphinx will read and
parse markdown files as well. This requires the `recommonmark` python
module which pulls in a lot of dependencies as well into the
requirements.txt file. You can read more about using Markdown with
Sphinx on readthedocs here:

https://docs.readthedocs.io/en/latest/intro/getting-started-with-sphinx.html#using-markdown-with-sphinx