Mounting private/data/archive to /ega/archive and updating the prefix handling for the POSIX FileStorage driver
volumes are created by docker-compose and owned by root.
Simply injecting a directory solves the issue for the moment,
because it's then owned by lega.

A subsequent PR will investigate the user of mount options for the
driver 'local'. I have tried `gid=lega,setgid=lega,noexec,nodev`,
but in this commit, I reverted to simplicity