Debugging on Travis and on Oscar's Linux host machine.
We experience that the docker driver for OSX might have a bug.
A root-owned directory should not be writable by the lega user, when permissions are 755.
But apparently, it is the case.

That implies that when the code, running under the lega user, creates subdirectories, it is not raising a PermissionError exception.
On Travis, it does.