Removed config assertion to enable config-file less config
The config lookup works by first checking for the <SECTION>_<VAR> env
var and then looking for the specified <var> in the [<section>] of the
config file. Checking for a specific section in the configuration then
will only work if there is a configuration file, since sections are not
a defined concept in the environment. The test is also unnecesary since
the following var lookup will fail anyway if it's not present.