Updating the code with Crypt4GH (latest version: 1)
The verify now uses a module to load the key.

We use a locked key on file, and unlock it with a passphrase.

Crypt4GH header is different, but still similiar: we store it in the DB entirely.
The bulk data portion is now encrypted with Chacha20Poly1305 and we do not have the built-in checksums.

Crypt4GH current version has built-in checksums _per_ block.
It is lacking an MDC field (Modification Detection Code), which basically is a checksum to validate that all the blocks are present and in the right order.
We either part from Crypt4GH and include it, or contribute to upgrade Crypt4GH.

Conflicts:
	lega/verify.py