Replacing pika with amqpstorm.
Pika was raising connection errors for long running jobs because
it could not fullfil the heartbeat _and_ the job concurrently.
The broker was then assuming that the connection was stale and
made the socket timeout.

This commit solves issue #63. The comment in #63 explains how to implement
a solution with pika, but after I tried it, I found the code very hard to read
and maintain.

After seeing the source code of AMQPStorm, I liked the idea that they spawn the
heartbeat in a separate thread, so that we don't need to resort to the callback trick
described in #63, and it is much cleaner.

I know that Python threads are not executing in parallel (cuz of the GIL), but that's not what I tried to
achieve. There is no real parallelism here. All I wanted is that the long
running jobs do not prevent the heartbeat ticking. Threads are just fine for that.