More robust connection from LocalEGA to CentralEGA
The shovel is now backed by a durable question instead of an ephemeral queue.
The advantages is that if Central EGA is down for a short time,
and a message gets shoveled at the time, the message is not lost, and it will be
delivered when Central EGA is back on again.

This mechanism is akin to a (reverse) federated queue, but lower level,
and we don't need credentials to the LocalEGA instance.